small fixes, bitwise debug output
git-svn-id: http://qon.jwdt.org/var/svn/catracker@200 97746f1a-7bc6-4ed5-8832-28bd89886fd8

+++ GPU/TPCCAGPUTracking/tracking-ca/standalone/code/AliHLTTPCCATracker.h
@@ -225,7 +225,7 @@ class AliHLTTPCCATracker
   GPUd() void MaximizeHitWeight( const AliHLTTPCCARow &row, int hitIndex, int weight ) {
     fData.MaximizeHitWeight( row, hitIndex, weight );
   }
-  GPUd() float HitWeight( const AliHLTTPCCARow &row, int hitIndex ) const {
+  GPUd() int HitWeight( const AliHLTTPCCARow &row, int hitIndex ) const {
     return fData.HitWeight( row, hitIndex );
   }
